refactor: search input as a standalone component
